requirements.txt
djangorestframework==3.5.3
Django==1.10.5